refactor(web): extract CalendarView, GalleryGrid, TimelineBar from Timelapse.svelte
